+++ src/test/groups.test.js
@@ -73,5 +73,28 @@ describe("Groups", function () {
 
       GroupsAction.deleteGroup("/group-non-existent");
     });
+
+    [403, 401, 404].forEach(status => {
+      it("emits an error on a non-existing group with statuscode " + status,
+        function (done) {
+          nock(config.apiURL)
+            .delete("/v2/groups//group-non-existent")
+            .reply(status, {
+              "message": "Group '/group-non-existent' does not exist"
+            });
+
+          GroupsStore.once(GroupsEvents.DELETE_ERROR,
+            function (error, statusCode) {
+              expectAsync(function () {
+
+                expect(statusCode)
+                  .to.equal(status);
+              }, done);
+            }
+          );
+
+          GroupsAction.deleteGroup("/group-non-existent");
+        });
+    });
   });
 });